[sandbox] rename App.tsx to Sandbox.tsx
