+++ packages/react-sandbox/src/Sandbox.test.tsx
@@ -1,9 +1,9 @@
 import React from 'react'
 import ReactDOM from 'react-dom'
-import App from './App'
+import Sandbox from './Sandbox'
 
 it('renders without crashing', () => {
   const div = document.createElement('div')
-  ReactDOM.render(<App />, div)
+  ReactDOM.render(<Sandbox />, div)
   ReactDOM.unmountComponentAtNode(div)
 })
